# Skip boxes that don't have actuators