--- a/amr-wind/wind_energy/actuator/ActSrcLineOp.H
+++ b/amr-wind/wind_energy/actuator/ActSrcLineOp.H
@@ -92,7 +92,7 @@ void ActSrcOp<ActTrait, ActSrcLine>::operator()(
 
     const auto& bx = mfi.tilebox();
 
-    const auto bxa = utils::realbox_to_box(actObj.m_data.info().bound_box, geom);
+    const auto bxa = utils::realbox_to_box(m_data.info().bound_box, geom);
     const auto& bxi = bx & bxa;
     if (bxi.isEmpty()) {
       return;