--- a/amr-wind/wind_energy/actuator/actuator_utils.cpp
+++ b/amr-wind/wind_energy/actuator/actuator_utils.cpp
@@ -3,15 +3,6 @@
 
 namespace amr_wind::actuator::utils {
 
-namespace {
-
-/** Convert a bounding box into amrex::Box index space at a given level
- *
- *  \param rbx Bounding box as defined in global domain coordinates
- *  \param geom AMReX geometry information for a given level
- *  \return The Box instance that defines the index space equivalent to bounding
- * boxt
- */
 amrex::Box
 realbox_to_box(const amrex::RealBox& rbx, const amrex::Geometry& geom)
 {
@@ -33,8 +24,6 @@ realbox_to_box(const amrex::RealBox& rbx, const amrex::Geometry& geom)
 
     return amrex::Box{lo, hi};
 }
-
-} // namespace
 
 std::set<int> determine_influenced_procs(
     const amrex::AmrCore& mesh, const amrex::RealBox& rbx)