--- a/amr-wind/wind_energy/actuator/disk/disk_spreading.H
+++ b/amr-wind/wind_energy/actuator/disk/disk_spreading.H
@@ -47,10 +47,11 @@ public:
     {
         const auto& bx = mfi.tilebox();
 
-        const auto bxa = utils::realbox_to_box(actObj.m_data.info().bound_box, geom);
+        const auto bxa =
+            utils::realbox_to_box(actObj.m_data.info().bound_box, geom);
         const auto& bxi = bx & bxa;
         if (bxi.isEmpty()) {
-          return;
+            return;
         }
 
         const auto& sarr = actObj.m_act_src(lev).array(mfi);
@@ -109,12 +110,13 @@ public:
     {
         const auto& bx = mfi.tilebox();
 
-        const auto bxa = utils::realbox_to_box(actObj.m_data.info().bound_box, geom);
+        const auto bxa =
+            utils::realbox_to_box(actObj.m_data.info().bound_box, geom);
         const auto& bxi = bx & bxa;
         if (bxi.isEmpty()) {
-          return;
+            return;
         }
-        
+
         const auto& sarr = actObj.m_act_src(lev).array(mfi);
         const auto& problo = geom.ProbLoArray();
         const auto& dx = geom.CellSizeArray();
@@ -173,12 +175,13 @@ public:
     {
         const auto& bx = mfi.tilebox();
 
-        const auto bxa = utils::realbox_to_box(actObj.m_data.info().bound_box, geom);
+        const auto bxa =
+            utils::realbox_to_box(actObj.m_data.info().bound_box, geom);
         const auto& bxi = bx & bxa;
         if (bxi.isEmpty()) {
-          return;
+            return;
         }
-        
+
         const auto& sarr = actObj.m_act_src(lev).array(mfi);
         const auto& problo = geom.ProbLoArray();
         const auto& dx = geom.CellSizeArray();