--- a/amr-wind/wind_energy/actuator/turbine/ActSrcDiskOp_Turbine.H
+++ b/amr-wind/wind_energy/actuator/turbine/ActSrcDiskOp_Turbine.H
@@ -84,6 +84,13 @@ operator()(const int lev, const amrex::MFIter& mfi, const amrex::Geometry& geom)
     BL_PROFILE("amr-wind::ActSrcOp<" + fname + ">");
 
     const auto& bx = mfi.tilebox();
+
+    const auto bxa = utils::realbox_to_box(m_data.info().bound_box, geom);
+    const auto& bxi = bx & bxa;
+    if (bxi.isEmpty()) {
+      return;
+    }
+
     const auto& sarr = m_act_src(lev).array(mfi);
     const auto& problo = geom.ProbLoArray();
     const auto& dx = geom.CellSizeArray();