@marchdf 2024-07-09
On a case with o(1000) turbines I am seeing:
`amr-wind::ActSrcOp<UniformCtDisk>                                       1342     0.6288      1.347      30.77   5.77%`
go down to this:
`amr-wind::ActSrcOp<UniformCtDisk>                                       1342    0.00011    0.02295      1.048   0.20%`
with this PR. 

If this works out on other cases, I will clean up the PR and push the same kind of fix to the other disks/spreading functions.