@mbkuhn 2024-07-10
do the changes in this PR also apply to post_regrid_actions and not just post init? I think they do, but I just wanted to check. You may want to add the profiling calls to that section too, while you're at it.