@marchdf 2024-07-10
Actually didn't mean to have those profiling calls still in there. Let me remove those. This is just for the src term calculation: skip all the boxes where the actuator won't have an influence on the src term.