@marchdf 2024-07-10
The questions I have for those more familiar with this code:

- I don't get diffs from the reg tests and some simple 2 turbine cases I ran. But can we be sure that checking the intersection of the `bound_box` with the mfiter box is enough?
- I would like to do the same kind of check for this one: https://github.com/Exawind/amr-wind/blob/main/amr-wind/wind_energy/actuator/turbine/ActSrcDiskOp_Turbine.H#L86. But it looks like it has a tower and hub. Does the `bound_box` include these? I couldn't figure out a sample case that uses this and where the bound box is computed for that. Does anyone know? @psakievich looks like you are on this blame but please redirect me if there is someone else who might know.