@mbkuhn 2024-07-10
I know from experience with 2D and moving fixed wing cases that if the expected force distribution exceeds the bounding box, it won't be there -- it won't get calculated on the mesh. So, I would think that the bound_box would have to cover everywhere that force is expected to be calculated.

Looking at where bound_box is calculated for different things:
- for standalone actuator disks (joukowsky, uniform_ct), there appears to be no tower representation in the reg tests, and when the bounding box is calculated in wind_energy/actuator/disk/disk_ops.cpp, only the disk itself is considered.
- for turbines (OpenFAST coupled, both disk and lines), the bounding box includes the tower (and then some) when it is calculated in wind_energy/actuator/turbine/turbine_utils.cpp. In z, the lower bound is the base position (already at the ground) - 1.25 * the radius. Probably should change that to just the base position, but it shouldn't be hurting anything the way it is.