@marchdf 2024-07-10
Ok cool, thanks for double checking that. And yes, the bound_box is originally used to check the procs that are influenced by a given actuator. So I would expect nothing to happen if the force extent is larger than the bound box extend. This PR extends that to to bx's used for the source term. So I think we are all good here and I will make sure this also happens for turbines.